Add base client to allow async overrides
